+++ apps/api-extractor/src/analyzer/ExportAnalyzer.ts
@@ -14,6 +14,7 @@ import { IFetchAstSymbolOptions } from './AstSymbolTable';
 import { AstEntity } from './AstEntity';
 import { AstNamespaceImport } from './AstNamespaceImport';
 import { SyntaxHelpers } from './SyntaxHelpers';
+import { last } from 'lodash';
 
 /**
  * Exposes the minimal APIs from AstSymbolTable that are needed by ExportAnalyzer.
@@ -444,9 +445,10 @@ export class ExportAnalyzer {
         //   import('api-extractor-lib1-test')
         //
         // Extracted qualifier:
-        //   apiExtractorLib1Test
+        //   ''
 
-        exportName = SyntaxHelpers.makeCamelCaseIdentifier(externalModulePath);
+        // exportName = SyntaxHelpers.makeCamelCaseIdentifier(externalModulePath);
+        exportName = '';
       }
 
       return this._fetchAstImport(undefined, {
@@ -752,6 +754,59 @@ export class ExportAnalyzer {
             isTypeOnly: false
           });
         }
+        // EqualsImport by namespace.
+        // EXAMPLE:
+        // import myLib2 = myLib;
+        // import B = myLib.A.B;
+      } else {
+        const reversedIdentifiers: ts.Identifier[] = [];
+        if (ts.isIdentifier(declaration.moduleReference)) {
+          reversedIdentifiers.push(declaration.moduleReference);
+        } else {
+          let current: ts.QualifiedName | undefined = declaration.moduleReference;
+          while (true) {
+            reversedIdentifiers.push(current.right);
+            if (ts.isIdentifier(current.left)) {
+              reversedIdentifiers.push(current.left);
+              break;
+            } else {
+              current = current.left;
+            }
+          }
+        }
+
+        const exportSubPath: string[] = [];
+        let externalImport: AstImport | undefined;
+        for (let i = reversedIdentifiers.length - 1; i >= 0; i--) {
+          const identifier: ts.Identifier = reversedIdentifiers[i];
+          if (!externalImport) {
+            // find the first external import as the base namespace
+            const symbol: ts.Symbol | undefined = this._typeChecker.getSymbolAtLocation(identifier);
+            if (!symbol) {
+              throw new Error('Symbol not found for identifier: ' + identifier.getText());
+            }
+            const astEntity: AstEntity | AstImport | undefined = this.fetchReferencedAstEntity(symbol, false);
+            if (astEntity instanceof AstImport) {
+              externalImport = astEntity;
+            }
+          } else {
+            exportSubPath.push(identifier.getText().trim());
+          }
+        }
+
+        if (externalImport) {
+          if (exportSubPath.length === 0) {
+            return externalImport;
+          } else {
+            return this._fetchAstImport(declarationSymbol, {
+              importKind: externalImport.importKind,
+              modulePath: externalImport.modulePath,
+              exportName: last(exportSubPath)!,
+              exportPath: externalImport.exportPath.concat(exportSubPath),
+              isTypeOnly: false
+            });
+          }
+        }
       }
     }
 
@@ -799,6 +854,25 @@ export class ExportAnalyzer {
   public tryGetExportOfAstModule(exportName: string, astModule: AstModule): AstEntity | undefined {
     const visitedAstModules: Set<AstModule> = new Set<AstModule>();
     return this._tryGetExportOfAstModule(exportName, astModule, visitedAstModules);
+  }
+
+  public tryGetReferencedAstImport(astImport: AstImport): AstImport | undefined {
+    if (astImport.exportPath) {
+      const referencedImport: AstImport | undefined = this._astImportsByKey.get(
+        AstImport.getKey({
+          importKind: astImport.importKind,
+          modulePath: astImport.modulePath,
+          exportName: astImport.exportPath[0],
+          isTypeOnly: false
+        })
+      );
+      if (referencedImport === undefined) {
+        throw new Error(
+          `For an AstImport of "EqualsImport" from namespace, there must have a referenced base AstImport.`
+        );
+      }
+      return referencedImport;
+    }
   }
 
   private _tryGetExportOfAstModule(